Extract parseEnvFile() to utils.js
