+++ src/utils.js
@@ -1,5 +1,7 @@
 import fs from 'bare-fs';
 import path from 'bare-path';
+import crypto from 'hypercore-crypto';
+import b4a from 'b4a';
 
 export function printAsciiArt() {
     console.log(
@@ -31,4 +33,24 @@ export function calculateDirectorySize(dirPath) {
 
     traverse(dirPath);
     return totalSize;
+}
+
+export async function parseEnvFile() {
+    if (!fs.existsSync('.env')) {
+        throw new Error('Generate .env first!');
+    }
+    const envFileContent = fs.readFileSync('.env', 'utf8');
+    const keyPair = {
+        publicKey: b4a.from(envFileContent.match(/PUBLIC_KEY=([0-9a-f]+)/)[1], 'hex'),
+        secretKey: b4a.from(envFileContent.match(/SECRET_KEY=([0-9a-f]+)/)[1], 'hex')
+    };
+    if (!crypto.validateKeyPair(keyPair)) {
+        throw new Error('Key pair not valid');
+    }
+    const peerSizeLimitMatch = envFileContent.match(/PEER_SIZE_LIMIT_MB=(\d+)/);
+    const peerSizeLimitMB = peerSizeLimitMatch ? parseInt(peerSizeLimitMatch[1]) : undefined;
+    if (peerSizeLimitMB) {
+        console.log(`Parsed peer size limit: ${peerSizeLimitMB}MB`);
+    }
+    return { keyPair, peerSizeLimitMB };
 }
